Introduce LatencyTimer operator.
LatencyTimer provides a operator which can wrap any Flow-Shape and measure how long each element took.
The operator the following shape
I -> O, TimedResult[O] when wrapping a Flow[I,O]

The TimedResult contains the wrapped flows result as well in case further logic is needed. Example: a http-call should be measured and the processing time for the external system should be added as a metric to a histogram. In this case only flows that returned at least a successfull result (not 500) should be added to the metric.

The operator does only support flows which are preserving the order, so no async should be done inside (only mapAsync).