AccumulateWhileUnchanged: stop pull() after onTimer()
If `onTimer()` is called when `buffer` is not empty, it will `push()`
results but the `pull(in)` request will not be canceled. If additional
upstream elements arrive with a different `nextState` which would cause
`pushResults()` to be called again before the downstream indicates demand
via `onPull()`, then `nextElement` would be dropped.

Avoid pushing again after push() from onTimer() until onPull() is called
again to indicate downstream demand. This avoids dropping elements when
pushResults() is called but `out` is not available.

In case a similar issue occurs again, add a `require()` check which will
throw if nextElement would be discarded.