Valve: `getMode` fails after the stage is stopped
Once the valve stage has been stopped, the future returned by `getMode`
used to never complete (an async callback is invoked with a `Promise`,
but the callback action is never process once the stage is completed).
By using `invokeWithFeedback` instead of `invoke`, we can be sure that
that future returned by `getMode` will either fail if the valve is
already stopped or else will be processed by the valve.

Fixes #119